1.4.0
This update has a large number of changes. First and foremost, the
shuffling functionality now has a "parallel" processing feature, which
speeds up randomization by a good amount-- even with large modlists, it
should only take 4-5 minutes. Getting all doors is also now faster, and
no longer needs to be re-called.

Worldspace connection testing and whitelisting has been considerably
redone. The main player-facing feature is the addition of [IsolatedExteriors]
and [IsolatedInteriors] in WorldRandomizer.ini. This reflects new
functionality to handle edge cases, such as exterior doors only usable
with certain other doors, and interior doors in similar cases. For example,
hoover dam's towers for exterior, or Vault34c for interior. For users that
are comfortable with xEdit, there is now a formlist, WRUnusableDoors, which
should store doors that are entirely unreachable by players.

An additional new feature is [iProblemSolverDoors], also in the ini file.
This option now allows users to customize how many doors are saved by the
Problem Solver (or set to -1 to disable it).

These are the main features, but there are also a number of bugfixes:

- Fixed and integrated Uncut Extra Collection causing script issues
- Fixed obscured doors not updating after a certain point
- Handled more cases of problematic non-warp doors (Thorn, Vault 3, etc.)
- Added the ability for NPCs to go through doors they try to use too many times
- Handled a number of door cases:
- Lucky 38 exterior, HELIOS One deck, McCarran interrogation cell,
- Ashton silo control station
- Fixed shuffle sometimes re-setting after reloading same save
- Properly navmeshed Green House
- There are likely other smaller fixes I'm not remembering
- Removed some unused scripts